cleanup missing variables; refactor PharoGsError

+++ importToGemStone.sh
@@ -2,7 +2,9 @@
 if [ ! -d ./output ]; then
   mkdir ./output
 fi
-rm output/*
+if [ -f output/Object.out ]; then
+  rm output/*
+fi
 topaz -l -q << EOF
 input importToGemStone.tpz
 EOF
